+++ sample_app/app/models/user.rb
@@ -22,7 +22,5 @@ class User < ApplicationRecord
     update_attribute(:remember_digest, nil)
   end
 
-  class << self
-    include UserConcern
-  end
+  include Authentication
 end
